+++ src/app/(homepage)/page.tsx
@@ -1,460 +1,23 @@
 "use client";
 
-import { useEffect, useMemo, useRef, useState } from "react";
 import { useRouter } from "next/navigation";
-import { motion, useScroll, useTransform, useInView } from "framer-motion";
-import MediaCarousel from "@/design-system/components/MediaCarousel";
-import MediaCard from "@/design-system/components/MediaCard";
 import { Overlay, OverlayMedia } from "@/design-system/components/MediaOverlay";
 import Box from "@/design-system/primitives/Box";
 import Text from "@/design-system/primitives/Text";
 import Image from "@/design-system/primitives/Image";
 import Button from "@/design-system/primitives/Button";
-import Link from "@/design-system/primitives/Link";
-import Divider from "@/design-system/primitives/Divider";
-import useWindowSize from "@/lib/hooks/useWindowSize";
-import { breakpoints } from "../../../tailwind.config";
-import { ChevronRight } from "lucide-react";
-
-// Integrated Parallax Divider and Newspaper Section
-function ParallaxDividerWithNewspaper({ router }: { router: ReturnType<typeof useRouter> }) {
-  const dividerRef = useRef<HTMLDivElement>(null);
-  const newspaperRef = useRef<HTMLDivElement>(null);
-
-  const { scrollYProgress: dividerScroll } = useScroll({
-    target: dividerRef,
-    offset: ["start end", "end start"],
-  });
-
-  const isInView = useInView(newspaperRef, { once: true, amount: 0 });
-
-  // Image moves within fixed container
-  const imageY = useTransform(dividerScroll, [0, 1], ["-20%", "20%"]);
-
-  return (
-    <div className="relative">
-      {/* Parallax Divider Background */}
-      <div ref={dividerRef} className="relative h-[400px] overflow-hidden laptop:h-[400px]">
-        <motion.div style={{ y: imageY }} className="absolute inset-0">
-          <div className="absolute -inset-[20%] bg-[url('/succulent.png')] bg-cover bg-center" />
-        </motion.div>
-      </div>
-
-      {/* Newspaper Section - Paper overlay */}
-      <motion.div
-        ref={newspaperRef}
-        initial={{ opacity: 0, y: 100, rotateX: 15 }}
-        animate={isInView ? { opacity: 1, y: 0, rotateX: 0 } : { opacity: 0, y: 100, rotateX: 15 }}
-        transition={{ duration: 0.8, ease: [0.4, 0, 0.2, 1] }}
-        style={{ perspective: "1000px" }}
-        className="relative -mt-32 mx-auto w-full max-w-6xl px-6 laptop:px-8"
-      >
-        {/* Paper container with defined width */}
-        <div className="relative bg-white shadow-2xl border border-black/10">
-          <Box className="relative px-4 laptop:px-16 py-16">
-            <Box className="flex flex-col items-start justify-between gap-4 laptop:flex-row laptop:items-end">
-              <Text size={48} className="mt-2 font-serif tracking-tight laptop:text-[56px]">
-                Recent Articles
-              </Text>
-
-              <Link
-                href="/articles"
-                className="group inline-flex items-center gap-2 border-b-2 border-black/20 pb-1 text-[14px] font-medium text-black/80 transition-all hover:border-black/40 hover:text-black"
-              >
-                View all articles
-                <ChevronRight className="h-4 w-4 transition-transform group-hover:translate-x-1" />
-              </Link>
-            </Box>
-
-            <Divider my={8} />
-
-            {/* Newspaper Layout - Multi-column */}
-            <Box className="grid gap-8 laptop:grid-cols-12">
-              {/* Main Feature Article - Left Column */}
-              <div className="laptop:col-span-8 space-y-8">
-                <OverlayMedia className="overflow-hidden border-2 border-black/10 shadow-lg">
-                  <Image src="/icy.png" alt="A dramatic solar eclipse" width="w-full" ratio={16 / 10} />
-                  <Overlay background="gradient-black">
-                    <Box className="flex h-full w-full items-end">
-                      <Box className="w-full px-8 pb-8 pt-32 laptop:px-10 laptop:pb-10 laptop:pt-40">
-                        <Text
-                          size={48}
-                          color="white"
-                          className="mt-4 font-serif tracking-tight max-laptop:text-[28px] max-sm:text-[28px]"
-                        >
-                          Siberian Permafrost
-                        </Text>
-                        <Text
-                          size={18}
-                          color="white"
-                          className="mt-3 max-w-2xl font-light leading-relaxed opacity-95 max-sm:text-[16px]"
-                        >
-                          A photo-led story about the science (and spectacle) behind ice—built to read like a print spread.
-                        </Text>
-                        <Box className="mt-6">
-                          <Button className="inline-flex" color="marigold" size="lg" onClick={() => router.push("/articles")}>
-                            Read article
-                          </Button>
-                        </Box>
-                      </Box>
-                    </Box>
-                  </Overlay>
-                </OverlayMedia>
-                <OverlayMedia className="overflow-hidden border-2 border-black/10 shadow-lg">
-                  <Image src="/succulent.png" alt="A dramatic solar eclipse" width="w-full" ratio={16 / 10} />
-                  <Overlay background="gradient-black">
-                    <Box className="flex h-full w-full items-end">
-                      <Box className="w-full px-8 pb-8 pt-32 laptop:px-10 laptop:pb-10 laptop:pt-40">
-                        <Text
-                          size={48}
-                          color="white"
-                          className="mt-4 font-serif tracking-tight max-laptop:text-[28px] max-sm:text-[28px]"
-                        >
-                          Chasing Totality
-                        </Text>
-                        <Text
-                          size={18}
-                          color="white"
-                          className="mt-3 max-w-2xl font-light leading-relaxed opacity-95 max-sm:text-[16px]"
-                        >
-                          A photo-led story about the science (and spectacle) behind eclipses—built to read like a print spread.
-                        </Text>
-                        <Box className="mt-6">
-                          <Button className="inline-flex" color="marigold" size="lg" onClick={() => router.push("/articles")}>
-                            Read article
-                          </Button>
-                        </Box>
-                      </Box>
-                    </Box>
-                  </Overlay>
-                </OverlayMedia>
-              </div>
-
-              {/* Sidebar Articles - Right Column */}
-              <div className="laptop:col-span-4">
-                <Box className="space-y-6 border-l-2 border-black/10 pl-6 laptop:pl-8">
-                  <MediaCard
-                    mediaType="image"
-                    imageProps={{ src: "/london.png", alt: "A misty city skyline" }}
-                    subtitle="Science + Society"
-                    title="Urban Heat Islands, Explained"
-                    description="How cities trap heat and what we can do about it—a deep dive into urban climate science."
-                    mediaDirection="top"
-                    size="sm"
-                    rounded="none"
-                    shadow="none"
-                    color="white"
-                    className="w-full max-w-none border-b border-black/10 pb-6"
-                  />
-
-                  <MediaCard
-                    mediaType="image"
-                    imageProps={{ src: "/moss.png", alt: "Green moss texture" }}
-                    subtitle="Research Spotlight"
-                    title="How Microbiomes Shape Our World"
-                    description="Exploring the invisible ecosystems that influence everything from our health to our environment."
-                    mediaDirection="top"
-                    size="sm"
-                    rounded="none"
-                    shadow="none"
-                    color="white"
-                    className="w-full max-w-none border-b border-black/10 pb-6"
-                  />
-                </Box>
-              </div>
-            </Box>
-
-            {/* Secondary Articles Grid - Newspaper Style */}
-            <div className="mt-12 border-t-2 border-black/10 pt-12">
-              <Box className="grid gap-6 laptop:grid-cols-3">
-                <MediaCard
-                  mediaType="image"
-                  imageProps={{ src: "/icy.png", alt: "Icy texture" }}
-                  subtitle="Quick Read"
-                  title="5 Questions About CRISPR"
-                  description="A concise guide to understanding gene editing technology."
-                  mediaDirection="top"
-                  size="sm"
-                  rounded="none"
-                  shadow="none"
-                  color="white"
-                  className="w-full max-w-none border border-black/10"
-                />
-                <MediaCard
-                  mediaType="image"
-                  imageProps={{ src: "/eclipse-image.png", alt: "Solar eclipse" }}
-                  subtitle="Opinion"
-                  title="Why Science Needs Better Stories"
-                  description="How narrative can bridge the gap between research and public understanding."
-                  mediaDirection="top"
-                  size="sm"
-                  rounded="none"
-                  shadow="none"
-                  color="white"
-                  className="w-full max-w-none border border-black/10"
-                />
-                <MediaCard
-                  mediaType="image"
-                  imageProps={{ src: "/logo.png", alt: "NU Sci logo mark" }}
-                  subtitle="Behind the Scenes"
-                  title="Designing an Issue Cover"
-                  description="A look into the creative process behind our magazine design."
-                  mediaDirection="top"
-                  size="sm"
-                  rounded="none"
-                  shadow="none"
-                  color="white"
-                  className="w-full max-w-none border border-black/10"
-                />
-              </Box>
-            </div>
-          </Box>
-        </div>
-      </motion.div>
-    </div>
-  );
-}
-
-function useCountUp(target: number, durationMs: number) {
-  const [value, setValue] = useState(0);
-
-  useEffect(() => {
-    const prefersReducedMotion =
-      typeof window !== "undefined" && window.matchMedia?.("(prefers-reduced-motion: reduce)")?.matches;
-
-    if (prefersReducedMotion || durationMs <= 0) {
-      setValue(target);
-      return;
-    }
-
-    let rafId = 0;
-    const start = performance.now();
-
-    const tick = (now: number) => {
-      const t = Math.min(1, (now - start) / durationMs);
-      const next = Math.round(target * (1 - Math.pow(1 - t, 3)));
-      setValue(next);
-      if (t < 1) rafId = requestAnimationFrame(tick);
-    };
-
-    rafId = requestAnimationFrame(tick);
-    return () => cancelAnimationFrame(rafId);
-  }, [target, durationMs]);
-
-  return value;
-}
+import Hero from "./components/Hero";
+import PrintMagazines from "./components/PrintMagazines";
+import FeaturedArticles from "./components/FeaturedArticles";
 
 export default function Homepage() {
   const router = useRouter();
-  const { width } = useWindowSize();
-
-  const issueThumbnails = useMemo(
-    () => [
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue60.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue58.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue56.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue54.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue52.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue50.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue48.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue46.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue44.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue42.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue40.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue38.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue36.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue34.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue32.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue30.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue28.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue26.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue24.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue22.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue20.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue18.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue16.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue14.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue12.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue10.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue8.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue6.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue4.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue2.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue1.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue3.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue5.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue7.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue9.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue11.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue13.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue15.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue17.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue19.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue21.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue23.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue25.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue27.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue29.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue31.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue33.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue35.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue37.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue39.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue41.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue43.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue45.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue47.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue49.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue51.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue53.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue55.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue57.png",
-      "https://northeasternsciencemagazine.github.io/nusci-issuu/thumbnails/issue59.png",
-    ],
-    []
-  );
-
-  const scrollToFeatured = () => {
-    document.getElementById("featured-issues")?.scrollIntoView({ behavior: "smooth", block: "start" });
-  };
-
-  const legacyYears = useCountUp(17, 1000);
-  const publishedIssues = useCountUp(66, 2000);
-  const reachArticles = useCountUp(1300, 3000);
 
   return (
     <main>
-      {/* HERO */}
-      <OverlayMedia className="w-full h-[620px] laptop:h-[700px] overflow-hidden">
-        {/* Fixed-height hero image: height stays consistent regardless of viewport width */}
-        <Image raw src="/moss.png" alt="Moss texture background" width="w-full" />
-        <Overlay background="solid-black">
-          <Box className="h-full w-full">
-            <Box className="mx-auto w-full max-w-6xl px-6 py-16 laptop:py-20">
-              <Text size={12} color="white" className="uppercase tracking-[0.35em] opacity-90">
-                Northeastern University&apos;s student-run science magazine
-              </Text>
-
-              <Box mt={4}>
-                <Text size={96} color="white" className="leading-none tracking-tight max-laptop:text-[72px] max-sm:text-[56px]">
-                  NU Sci
-                </Text>
-                <Text size={30} color="white" className="mt-4 max-w-2xl font-light opacity-95 max-sm:text-[20px]">
-                  Science communication for the community. Written, designed, photographed, and developed by students.
-                </Text>
-              </Box>
-
-              <Box mt={10} className="flex flex-wrap gap-3">
-                <Button className="inline-flex" color="aqua" size="lg" onClick={scrollToFeatured}>
-                  Browse articles
-                </Button>
-                <Button className="inline-flex" variant="outline" color="white" size="lg" onClick={() => router.push("/")}>
-                  Learn more
-                </Button>
-              </Box>
-
-              {/* Stats Badges */}
-              <Box mt={10} className="grid grid-cols-3 gap-3 max-sm:grid-cols-1 mobile:invisible laptop:visible">
-                <Box className="rounded-2xl bg-white/10 backdrop-blur" px={4} py={4}>
-                  <Text size={12} color="white" className="uppercase tracking-[0.25em] opacity-80">
-                    Legacy
-                  </Text>
-                  <Box className="mt-1 flex items-baseline gap-2">
-                    <Text
-                      as="span"
-                      size={24}
-                      color="white"
-                      className="inline-block min-w-[2ch] text-right font-semibold tabular-nums"
-                    >
-                      {legacyYears.toLocaleString()}
-                    </Text>
-                    <Text as="span" size={24} color="white">
-                      Years
-                    </Text>
-                  </Box>
-                </Box>
-                <Box className="rounded-2xl bg-white/10 backdrop-blur" px={4} py={4}>
-                  <Text size={12} color="white" className="uppercase tracking-[0.25em] opacity-80">
-                    Published
-                  </Text>
-                  <Box className="mt-1 flex items-baseline gap-2">
-                    <Text
-                      as="span"
-                      size={24}
-                      color="white"
-                      className="inline-block min-w-[2ch] text-right font-semibold tabular-nums"
-                    >
-                      {publishedIssues.toLocaleString()}
-                    </Text>
-                    <Text as="span" size={24} color="white">
-                      Issues
-                    </Text>
-                  </Box>
-                </Box>
-                <Box className="rounded-2xl bg-white/10 backdrop-blur" px={4} py={4}>
-                  <Text size={12} color="white" className="uppercase tracking-[0.25em] opacity-80">
-                    Reach
-                  </Text>
-                  <Box className="mt-1 flex items-baseline gap-2">
-                    <Text
-                      as="span"
-                      size={24}
-                      color="white"
-                      className="inline-block min-w-[5.5ch] text-right font-semibold tabular-nums"
-                    >
-                      {reachArticles.toLocaleString()}+
-                    </Text>
-                    <Text as="span" size={24} color="white">
-                      Articles
-                    </Text>
-                  </Box>
-                </Box>
-              </Box>
-            </Box>
-          </Box>
-        </Overlay>
-      </OverlayMedia>
-
-      {/* FEATURED ISSUES (carousel is first featured item) */}
-      <Box id="featured-issues" className="scroll-mt-24 bg-white">
-        <Box className="mx-auto w-full max-w-6xl px-6 pt-14 pb-10">
-          <Box className="flex flex-col items-start justify-between gap-6 laptop:flex-row laptop:items-end">
-            <Box>
-              <Text size={36} className="tracking-tight">
-                Our Magazines
-              </Text>
-              <Text size={16} className="mt-2 max-w-2xl text-black/70">
-                Look through our print archive - click a cover to bring it front and center.
-              </Text>
-            </Box>
-
-            <Box className="flex items-center gap-3">
-              <Link
-                href="https://northeasternsciencemagazine.github.io/nusci-issuu/"
-                newWindow
-                className="rounded-full border border-black/15 px-4 py-2 text-[14px] text-black/80 hover:bg-black/5"
-              >
-                View the archive
-              </Link>
-            </Box>
-          </Box>
-
-          <Divider mt={8} />
-
-          {/* Technically responsive - renders w/ different props at breakpoint laptop */}
-          <MediaCarousel
-            media={issueThumbnails}
-            size={width && width > breakpoints.laptop ? "lg" : "md"}
-            visibleCount={width && width > breakpoints.laptop ? 7 : 3}
-            initialIndex={0}
-          />
-
-          <Divider mt={8} />
-        </Box>
-      </Box>
-
-      {/* Parallax Divider with Newspaper Overlay */}
-      <ParallaxDividerWithNewspaper router={router} />
+      <Hero />
+      <PrintMagazines />
+      <FeaturedArticles />
 
       {/* CTA OVERLAY */}
       <Box className="bg-white py-16">
